Feature: added boolean required validator.
This validator adds when the field is `required` and a `boolean`.
The behavior is that the user must then set the `boolean` to `true`
otherwise the field is invalid.

Also fixed flaky validator tests which used `resolves` but did not
return the `Promise`. This causes tests to silently fail, but succeed
in the CI. Rewritten to use `async` with `done` instead