improvement: boolean required now accepts `false` as well as `true`.
Before this commit the `makeBooleanRequired` required the boolean to
be of value `true`.

This is strange in the following situation: lets say we have a dropdown
which shows three options and their values:

1. "Do you agree", undefined
2. "Yes", true
3. "No" false

When you had a `makeBooleanRequired` we currently only consider the
option "Yes" valid. This is strange because "No" is perfectly valid.
The only invalid choose is "Do you agree" because it is undefined.

Historically a boolean required was considered sort of a "agree"
choice which the user had to answer `true` too.We must change this
behavior to support the other use cases.

One caveat is that for a checkbox we would not accept the `undefined`
value as valid. This means that the developer must set the default values
of all checkboxes too `false` manually. Which is a slight burden but
this way we support all use cases.

Also changed the `package.json` scripts to the new script standard.
Silenced all `eslint` warnings of our very dynamic code as a result
of this.

Closes: #7